Update core-functions-in-depth.md
This might be overly pedantic, but I don't think so: append means, especially in computer-sciencey terms, "Add to the end."

I found it confusing that you were asking me to append to the suspect __list__. If we wanted to grow this data at the tail, shouldn't we have used a vector? Were you just asking me to do this inefficient thing as an exercise? Then I noticed that you had described "appending to the beginning" earlier in the chapter.

Anyway, thanks a ton for the awesome resource. Been super helpful in getting my feet wet with Clojure!